email-service/requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
pip==24.2
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.13
wheel==0.43.0
kombu>=5.3.4,<6.0.0
mysqlclient==2.2.0
gunicorn==20.1.0
celery==5.3.6
pika